Fix IP_TYPE import
